Move logic for object deletion to space plugin root
